requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
django-allauth==0.63.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.0
tzdata==2024.1
